refactor text_to_speech.
