+++ src/cpp/src/module_genai/modules/md_text_to_speech/md_text_to_speech.hpp
@@ -3,32 +3,42 @@
 
 #pragma once
 
-#ifdef ENABLE_OPENVINO_NEW_ARCH
-
 #include "module_genai/pipeline/module.hpp"
 #include "module_genai/pipeline/pipeline_impl.hpp"
-#include "openvino/genai/tokenizer.hpp"
-#include "modeling/models/qwen3_omni/processing_qwen3_omni.hpp"
+
 #include <memory>
-#include <vector>
-#include <optional>
-#include <string>
 #include <filesystem>
-#include <variant>
+#include <optional>
 #include <random>
+#include <string>
+#include <vector>
+
+#include "openvino/genai/tokenizer.hpp"
 
 namespace ov::genai::module {
 
-using Qwen3OmniProcessingConfig = modeling::models::Qwen3OmniProcessingConfig;
-
 class TextToSpeechModule : public IBaseModule {
-    DeclareModuleConstructor(TextToSpeechModule)
+protected:
+    TextToSpeechModule() = delete;
+    TextToSpeechModule(const IBaseModuleDesc::PTR& desc,
+                       const PipelineDesc::PTR& pipeline_desc,
+                       const VLMModelType& model_type);
 
-private:
-    bool initialize();
+public:
+    virtual ~TextToSpeechModule();
+
+    using PTR = std::shared_ptr<TextToSpeechModule>;
+    static PTR create(const IBaseModuleDesc::PTR& desc, const PipelineDesc::PTR& pipeline_desc);
+    static void print_static_config();
+    void run() override;
+
+protected:
+    std::vector<std::string> parse_input_texts();
+
     std::optional<std::filesystem::path> get_model_path(const std::string &param_name);
-    std::shared_ptr<ov::Model> load_model(const std::filesystem::path &model_path, const std::string &default_model_filename = "");
-    std::pair<ov::Tensor, int> qwen3_omni_text_to_speech(const std::string &text);
+    std::shared_ptr<ov::Model> load_model(const std::filesystem::path &model_path,
+                                          const std::string &default_model_filename = "");
+
     std::vector<int64_t> make_mrope_positions(size_t start, size_t len, size_t batch);
     ov::Tensor make_causal_mask(size_t seq_len, size_t batch);
     int64_t sample_codec_token(const float* logits, size_t vocab_size,
@@ -40,6 +50,10 @@ private:
     ov::Tensor make_decode_mask(size_t past_len, size_t batch);
     std::pair<ov::Tensor, int> synthesize_fallback_tone(const std::string& text);
 
+protected:
+    VLMModelType m_model_type;
+    std::string m_device = "CPU";
+
     std::unique_ptr<ov::InferRequest> m_embedding_infer;
     std::unique_ptr<ov::InferRequest> m_prefill_infer;
     std::unique_ptr<ov::InferRequest> m_decode_infer;
@@ -49,9 +63,6 @@ private:
     std::unique_ptr<ov::InferRequest> m_code_predictor_single_codec_embedding_infer;
     std::unique_ptr<ov::InferRequest> m_speech_decoder_infer;
     std::unique_ptr<Tokenizer> m_tokenizer;
-    std::variant<Qwen3OmniProcessingConfig> m_config;
 };
 
 }
-
-#endif
